fix(worker): remove trailing '=' from working dir
Why : some makefile can try to interprete '=' in path.
ex : OBJS = $(SRCS:$(SRCS_DIR)/%.c=$(OBJDIR)/%.o)
if SRCS_DIR contains '=' this fail to create the list of objs
Even if it should be up to the build script to manage this.
This commit will prevent this kind of tricky issue.

How : use RawStdEncoding instead of StdEncoding
RawStdEncoding is the standard raw, unpadded base64 encod